AI/week2/projects/Sentiment Vision/requirements.txt
pandas
numpy
streamlit
tensorflow
pillow
fastai